refactor: Relocate PriceItemUiModel (#64)
