+++ app/src/main/java/com/example/rentit/presentation/rentaldetail/common/PriceItemUiModel.kt
@@ -1,4 +1,4 @@
-package com.example.rentit.presentation.rentaldetail.common.model
+package com.example.rentit.presentation.rentaldetail.common
 
 data class PriceItemUiModel(
     val label: String,
