+++ app/src/main/java/com/example/rentit/presentation/rentaldetail/common/components/PriceSummary.kt
@@ -19,7 +19,7 @@ import com.example.rentit.common.component.CommonDivider
 import com.example.rentit.common.theme.PrimaryBlue500
 import com.example.rentit.common.theme.RentItTheme
 import com.example.rentit.common.util.formatPrice
-import com.example.rentit.presentation.rentaldetail.common.model.PriceItemUiModel
+import com.example.rentit.presentation.rentaldetail.common.PriceItemUiModel
 
 /**
  * 비용 내역과 총합을 나타내는 UI 컴포넌트
